requirements.txt
Flask>=2.2.3
requests>=2.7.0
google-cloud-datastore>=2.1.6